refactor dev api integrations to standalone classes
