+++ checkov/common/bridgecrew/integration_features/__init__.py
@@ -1,0 +1,1 @@
+from checkov.common.bridgecrew.integration_features.features import *
